Update_backEnd

+++ SWP490_TW/SWP490_TW/src/main/java/com/capstone/project/domain/Order.java
@@ -23,7 +23,7 @@ public class Order {
     private User user;
 
     @Column(name = "total_price")
-    private float totalPrice;
+    private Double totalPrice;
 
     @Column(name = "note")
     private String note;
