+++ SWP490_TW/SWP490_TW/src/main/java/com/capstone/project/request/OrderDetailRequest.java
@@ -15,6 +15,6 @@ public class OrderDetailRequest {
     private int orderId;
     private int productId;
     private String productName;
-    private float productPrice;
+    private Double productPrice;
     private int quantity;
 }
