+++ SWP490_TW/SWP490_TW/src/main/java/com/capstone/project/request/ProductRequest.java
@@ -21,9 +21,9 @@ public class ProductRequest {
     private int productId;
     private String productName;
     private String description;
-    private float originalPrice;
-    private float salePercent;
-    private float sellPrice;
+    private Double originalPrice;
+    private Double salePercent;
+    private Double sellPrice;
     private int amount;
     private Date createdDate;
     private int productStatus;
