+++ SWP490_TW/SWP490_TW/src/main/java/com/capstone/project/response/OrderDetailResponse.java
@@ -16,6 +16,6 @@ public class OrderDetailResponse {
     private Order order;
     private Product product;
     private String productName;
-    private float productPrice;
+    private Double productPrice;
     private int quantity;
 }
