+++ SWP490_TW/SWP490_TW/src/main/java/com/capstone/project/response/PromotionResponse.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class PromotionResponse {
     private int promotionId;
     private String promotionCode;
-    private float salePercent;
+    private Double salePercent;
     private int amount;
 }
